Applied GenAI/Solutions/multi-type-parsing-and-vector-search-2/pyproject.toml
[project]
name = "multi-type-parsing-and-vector-search-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.3"
dependencies = [
    "google-cloud-bigquery>=3.40.1",
    "google-cloud-bigquery-connection>=1.20.0",
    "google-cloud-documentai>=3.10.0",
    "google-cloud-storage>=3.9.0",
    "google-cloud-vectorsearch>=0.5.0",
    "google-genai>=1.64.0",
    "langextract>=0.1.0",
    "ipykernel>=7.2.0",
    "nbconvert>=7.17.0",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "pydantic>=2.12.5",
    "weasyprint>=68.1",
]
